feat(movie): add movie and movies queries
Add `movie` and `movies` queries to retrieve a single movie and a list of movies, respectively. The `movies` query accepts a `FindMoviesInput` input type to filter the results by title. The `movies` query also uses the `PaginationArgs` input type to enable pagination. The `MovieService` class was updated to implement these queries. A `MovieMapper` class was also added to map the Prisma entities to the `MovieModel` class.

feat(schema.graphql): add new types and fields for movies, collections, images, media, persons, and more