feat(movie.graphql, schema.graphql): add support for country model and countryId field in ImageLinkModel
This commit adds support for the CountryModel in the schema and a new query to retrieve a single country by ID. Additionally, the ImageLinkModel now includes a new field, countryId, to reference the country associated with the image. The movie.graphql file has been updated to include a new query to retrieve genres and their associated images.