feat(page-info): add update mutation to PageInfoResolver and PageInfoService
Add an update mutation to PageInfoResolver and PageInfoService to allow updating of PageInfo entities. The mutation accepts an UpdatePageInfoInput object as an argument and returns a PageInfoModel object. The PageInfoService's update method updates the PageInfo entity with the given id and data. The defaultIncludes property is added to the PageInfoService to include the seo and image entities with their respective assets.