feat(season): add createMany and updateMany mutations
Add `createMany` and `updateMany` mutations to the `SeasonResolver` to create and update multiple seasons at once. The `createMany` mutation takes an array of `CreateSeasonInput` objects and returns an array of `SeasonModel` objects. The `updateMany` mutation takes an array of `UpdateSeasonInput` objects and returns an array of `SeasonModel` objects. The `SeasonService` has been updated to handle these mutations.