feat(image-processing.service.ts): add downloadImage method to ImageProcessingService
The `downloadImage` method is added to the `ImageProcessingService` class. This method takes a URL as input and returns a Promise that resolves to a Buffer containing the downloaded image data.

The method uses the `fetch` function to make a GET request to the specified URL. If the response is not successful (status code other than 200), an error is thrown with the response status text.

If the response is successful, the image data is retrieved as an ArrayBuffer using the `arrayBuffer` method of the response. The ArrayBuffer is then converted to a Buffer using the `Buffer.from` method.

The resulting Buffer is returned as the result of the Promise.